[#13] 디자인 수정대로 밈 생성 UI 변경
PreviewFragment에서 RecyclerView 아이템 생성 및 연결 필요
