+++ Android/app/src/main/java/com/snowball/memetory/presentation/ui/generatememe/PreviewDetailFragment.kt
@@ -1,0 +1,34 @@
+package com.snowball.memetory.presentation.ui.generatememe
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.databinding.DataBindingUtil
+import androidx.navigation.NavController
+import androidx.navigation.Navigation
+import com.snowball.memetory.R
+import com.snowball.memetory.databinding.FragmentPreviewDetailBinding
+
+class PreviewDetailFragment : Fragment() {
+
+    private lateinit var binding: FragmentPreviewDetailBinding
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        binding = DataBindingUtil.inflate(inflater, R.layout.fragment_preview_detail, container, false)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+    }
+
+}
