+++ Android/app/src/main/java/com/snowball/memetory/presentation/ui/generatememe/PreviewFragment.kt
@@ -8,16 +8,17 @@ import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
 import androidx.navigation.NavController
 import androidx.navigation.Navigation
+import androidx.navigation.fragment.findNavController
 import com.snowball.memetory.R
-import com.snowball.memetory.databinding.FragmentEnterLinesBinding
+import com.snowball.memetory.databinding.FragmentPreviewBinding
 
-class EnterLinesFragment : Fragment() {
+class PreviewFragment : Fragment() {
 
     lateinit var navController: NavController
-    lateinit var binding: FragmentEnterLinesBinding
+    lateinit var binding: FragmentPreviewBinding
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-
     }
 
     override fun onCreateView(
@@ -25,7 +26,7 @@ class EnterLinesFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        binding = DataBindingUtil.inflate(inflater, R.layout.fragment_enter_lines, container, false)
+        binding = DataBindingUtil.inflate(inflater, R.layout.fragment_preview, container, false)
         return binding.root
     }
 
@@ -33,8 +34,11 @@ class EnterLinesFragment : Fragment() {
         super.onViewCreated(view, savedInstanceState)
 
         navController = Navigation.findNavController(view)
+        // 추후에 setOnClickListener를 confirmBtn이 아니라 리사이클러뷰에 있는 아이템으로 변경해야함.
+        // 아직 아이템 만들지 않고 연결을 안했음.
         binding.confirmBtn.setOnClickListener {
-            navController.navigate(R.id.action_enterLinesFragment_to_sceneDetailFragment)
+            navController.navigate(R.id.action_previewFragment_to_previewDetailFragment)
         }
     }
+
 }
