+++ Android/app/src/main/res/font/font.xml
@@ -1,5 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
 <font-family xmlns:android="http://schemas.android.com/apk/res/android">
-    <font android:font="@font/bmjua_otf" />
     <font android:font="@font/pretendard_bold"/>
 </font-family>
