--- Android/app/src/main/res/layout/fragment_choose_face.xml
@@ -1,75 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <data>
-
-    </data>
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:paddingVertical="20dp"
-        tools:context=".presentation.ui.generatememe.ChooseFaceFragment">
-
-        <androidx.appcompat.widget.AppCompatTextView
-            android:id="@+id/titleText"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:textAppearance="@style/TextAppearance.App.h3"
-            android:text="@string/scene_choose_face"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintEnd_toEndOf="parent" />
-
-        <androidx.constraintlayout.widget.ConstraintLayout
-            android:id="@+id/imgConstraintLayout"
-            android:layout_width="300dp"
-            android:layout_height="300dp"
-            android:background="@drawable/corner_radius_10"
-            android:backgroundTint="@color/subBtn"
-            android:layout_marginTop="40dp"
-            app:layout_constraintTop_toBottomOf="@id/titleText"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintEnd_toEndOf="parent">
-
-            <androidx.appcompat.widget.AppCompatImageView
-                android:id="@+id/chooseFaceImg"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:scaleType="centerCrop"
-                android:src="@drawable/ic_add_50"
-                app:layout_constraintBottom_toBottomOf="parent"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toTopOf="parent" />
-        </androidx.constraintlayout.widget.ConstraintLayout>
-
-        <androidx.appcompat.widget.AppCompatTextView
-            android:id="@+id/alertText"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:textAppearance="@style/TextAppearance.App.body3"
-            android:textColor="@android:color/darker_gray"
-            android:text="한 사람의 얼굴이 정 가운데에 위치한 사진을 선택해주세요!"
-            android:layout_marginTop="10dp"
-            app:layout_constraintStart_toStartOf="@id/imgConstraintLayout"
-            app:layout_constraintTop_toBottomOf="@id/imgConstraintLayout" />
-
-        <androidx.appcompat.widget.AppCompatButton
-            android:id="@+id/confirmBtn"
-            android:layout_width="300dp"
-            android:layout_height="wrap_content"
-            android:background="@drawable/corner_radius_10"
-            android:backgroundTint="@color/mainBtn"
-            android:textAppearance="@style/TextAppearance.App.button"
-            android:textColor="@color/white"
-            android:text="선택하기"
-            app:layout_constraintTop_toBottomOf="@id/imgConstraintLayout"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintEnd_toEndOf="parent" />
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-</layout>
