+++ Android/app/src/main/res/layout/fragment_choose_template.xml
@@ -34,7 +34,7 @@
             app:layout_constraintTop_toBottomOf="@id/titleText" />
 
         <androidx.appcompat.widget.AppCompatButton
-            android:id="@+id/confirmBtn"
+            android:id="@+id/chooseBtn"
             android:layout_width="300dp"
             android:layout_height="wrap_content"
             android:background="@drawable/corner_radius_10"
@@ -42,6 +42,20 @@
             android:textAppearance="@style/TextAppearance.App.button"
             android:textColor="@color/white"
             android:text="선택하기"
+            android:layout_marginBottom="20dp"
+            app:layout_constraintBottom_toTopOf="@+id/confirmBtn"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent" />
+
+        <androidx.appcompat.widget.AppCompatButton
+            android:id="@+id/confirmBtn"
+            android:layout_width="300dp"
+            android:layout_height="wrap_content"
+            android:background="@drawable/corner_radius_10"
+            android:backgroundTint="@color/mainBtn"
+            android:textAppearance="@style/TextAppearance.App.button"
+            android:textColor="@color/white"
+            android:text="@string/button_finish"
             app:layout_constraintTop_toBottomOf="@id/templateViewPager"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintStart_toStartOf="parent"
