+++ Android/app/src/main/res/layout/fragment_preview.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
+    xmlns:tools="http://schemas.android.com/tools"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
 
     <data>
 
@@ -11,33 +11,42 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:paddingTop="20dp"
-        tools:context=".presentation.ui.generatememe.EnterLinesFragment">
+        tools:context=".presentation.ui.generatememe.PreviewFragment">
 
         <androidx.appcompat.widget.AppCompatTextView
             android:id="@+id/titleText"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="@string/scene_enter_lines"
             android:textAppearance="@style/TextAppearance.App.h3"
-            app:layout_constraintEnd_toEndOf="parent"
+            android:text="확인하기"
+            app:layout_constraintTop_toTopOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
+            app:layout_constraintEnd_toEndOf="parent" />
 
         <androidx.appcompat.widget.AppCompatEditText
-            android:id="@+id/linesEditText"
+            android:id="@+id/titleEditText"
             android:layout_width="match_parent"
-            android:layout_height="300dp"
-            android:layout_marginHorizontal="30dp"
+            android:layout_height="wrap_content"
+            android:layout_marginHorizontal="20dp"
             android:layout_marginTop="30dp"
             android:background="@drawable/enter_lines_edittext_background"
             android:gravity="top|start"
-            android:hint="대사를 입력해주세요."
+            android:hint="제목"
             android:textAppearance="@style/TextAppearance.App.body2"
             android:padding="10dp"
             android:inputType="text"
-            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintTop_toBottomOf="@id/titleText"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/titleText" />
+            app:layout_constraintEnd_toEndOf="parent" />
+
+        <androidx.recyclerview.widget.RecyclerView
+            android:id="@+id/previewRecyclerView"
+            android:layout_width="match_parent"
+            android:layout_height="350dp"
+            android:layout_marginTop="40dp"
+            app:layout_constraintTop_toBottomOf="@id/titleEditText"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"/>
 
         <androidx.appcompat.widget.AppCompatButton
             android:id="@+id/confirmBtn"
@@ -45,13 +54,14 @@
             android:layout_height="wrap_content"
             android:background="@drawable/corner_radius_10"
             android:backgroundTint="@color/mainBtn"
-            android:text="@string/button_confirm"
             android:textAppearance="@style/TextAppearance.App.button"
             android:textColor="@color/white"
+            android:text="@string/button_finish"
+            android:layout_marginTop="30dp"
+            app:layout_constraintTop_toBottomOf="@id/previewRecyclerView"
             app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/linesEditText" />
+            app:layout_constraintEnd_toEndOf="parent" />
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
