+++ Android/app/src/main/res/layout/fragment_preview_detail.xml
@@ -1,0 +1,112 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+
+    <data>
+
+    </data>
+
+    <androidx.core.widget.NestedScrollView
+        android:layout_width="match_parent"
+        android:layout_height="match_parent">
+
+        <androidx.constraintlayout.widget.ConstraintLayout
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:paddingVertical="20dp"
+            tools:context=".presentation.ui.generatememe.PreviewDetailFragment">
+
+            <androidx.appcompat.widget.AppCompatTextView
+                android:id="@+id/titleText"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:textAppearance="@style/TextAppearance.App.h3"
+                android:text="확인하기"
+                app:layout_constraintTop_toTopOf="parent"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintEnd_toEndOf="parent" />
+
+            <androidx.appcompat.widget.AppCompatTextView
+                android:id="@+id/faceTitleText"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:textAppearance="@style/TextAppearance.App.h4"
+                android:text="얼굴"
+                android:layout_marginTop="20dp"
+                app:layout_constraintTop_toBottomOf="@id/titleText"
+                app:layout_constraintStart_toStartOf="@id/imgConstraintLayout" />
+
+            <androidx.constraintlayout.widget.ConstraintLayout
+                android:id="@+id/imgConstraintLayout"
+                android:layout_width="350dp"
+                android:layout_height="300dp"
+                android:background="@drawable/corner_radius_10"
+                android:backgroundTint="@color/subBtn"
+                app:layout_constraintTop_toBottomOf="@id/faceTitleText"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintEnd_toEndOf="parent">
+
+                <androidx.appcompat.widget.AppCompatImageView
+                    android:id="@+id/faceImg"
+                    android:layout_width="300dp"
+                    android:layout_height="300dp"
+                    android:scaleType="centerCrop"
+                    app:layout_constraintBottom_toBottomOf="parent"
+                    app:layout_constraintEnd_toEndOf="parent"
+                    app:layout_constraintStart_toStartOf="parent"
+                    app:layout_constraintTop_toTopOf="parent" />
+            </androidx.constraintlayout.widget.ConstraintLayout>
+
+            <androidx.appcompat.widget.AppCompatTextView
+                android:id="@+id/linesTitleText"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:textAppearance="@style/TextAppearance.App.h4"
+                android:text="대사"
+                android:layout_marginTop="20dp"
+                app:layout_constraintTop_toBottomOf="@id/imgConstraintLayout"
+                app:layout_constraintStart_toStartOf="@id/imgConstraintLayout" />
+
+            <androidx.constraintlayout.widget.ConstraintLayout
+                android:id="@+id/linesConstraintLayout"
+                android:layout_width="350dp"
+                android:layout_height="100dp"
+                android:background="@drawable/corner_radius_10"
+                android:backgroundTint="@color/subBtn"
+                android:padding="10dp"
+                app:layout_constraintTop_toBottomOf="@id/linesTitleText"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintEnd_toEndOf="parent">
+
+                <androidx.appcompat.widget.AppCompatTextView
+                    android:id="@+id/linesText"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:textAppearance="@style/TextAppearance.App.body2"
+                    android:text="대사"
+                    app:layout_constraintTop_toTopOf="@id/linesConstraintLayout"
+                    app:layout_constraintStart_toStartOf="@id/linesConstraintLayout" />
+            </androidx.constraintlayout.widget.ConstraintLayout>
+
+            <androidx.appcompat.widget.AppCompatTextView
+                android:id="@+id/voiceTitleText"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:textAppearance="@style/TextAppearance.App.h4"
+                android:text="목소리"
+                android:layout_marginTop="20dp"
+                app:layout_constraintTop_toBottomOf="@id/linesConstraintLayout"
+                app:layout_constraintStart_toStartOf="@id/linesConstraintLayout" />
+
+            <include
+                layout="@layout/item_choose_voice"
+                android:id="@+id/myIncludedView"
+                android:layout_width="350dp"
+                android:layout_height="wrap_content"
+                app:layout_constraintTop_toBottomOf="@id/voiceTitleText"
+                app:layout_constraintStart_toStartOf="@id/voiceTitleText" />
+        </androidx.constraintlayout.widget.ConstraintLayout>
+    </androidx.core.widget.NestedScrollView>
+
+</layout>
